mcp-trader/pyproject.toml
[project]
name = "mcp-trader"
version = "0.3.0"
description = "A sample MCP server for traders with technical analysis capabilities"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.12.13",
    "fastmcp>=2.8.0",
    "numpy~=1.26.4",
    "pandas>=2.3.0",
    "python-dotenv>=1.0.1",
    "pandas-ta >=0.3.14b0",
    "yfinance>=0.2.63",
    "setuptools>=80.9.0",
    "textblob>=0.19.0",
    "vadersentiment>=3.3.2",
    "scikit-learn>=1.3.0",
    "pydantic>=2.0.0", 
]
keywords = ["mcp", "trading", "technical-analysis", "finance","stocks", "crypto", "sentiment-analysis"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "License :: OSI Approved :: MIT License",
    "Topic :: Office/Business :: Financial :: Investment",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Intended Audience :: Financial and Insurance Industry",
    "Intended Audience :: Developers",
]

[project.urls]
Homepage = "https://github.com/oumyL/mcp-trader"
Repository = "https://github.com/oumyL/mcp-trader"
Issues = "https://github.com/oumyL/mcp-trader/issues"
Documentation = "https://github.com/oumyL/mcp-trader#readme"

[[project.authors]]
name = "Oumaima Tayabi"
email = "tayabi.ouma@gmail.com"

[project.optional-dependencies]
http = ["uvicorn>=0.34.3"]
redis = ["redis>=6.2.0"]
ml = [
    "xgboost>=2.0.0",
    "lightgbm>=4.0.0",
    "joblib>=1.3.0",
]
[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "ruff>=0.11.13",
    "mypy>=1.16.0",
    "types-aiofiles",
    "types-redis",
]
test = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
mcp-trader = "mcp_trader.server:main"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/mcp_trader"]

[tool.uv]
constraint-dependencies = [
    # Ensure NumPy 1.x for compatibility
    "numpy<2",
]

[tool.ruff]
line-length = 100
target-version = "py311"
src = ["mcp_trader"]
[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "SIM",  # flake8-simplify
    "RUF", # Ruff-specific rules
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "B904",  # raise without from in exception handler
]
[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]  # Allow unused imports in __init__.py
"test_*.py" = ["SIM117"]  # Allow multiple with statements in tests

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true
show_error_codes = true
# Ignore missing imports for optional dependencies
[[tool.mypy.overrides]]
module = [
    "pandas_ta.*",
    "yfinance.*",
    "textblob.*",
    "vaderSentiment.*",
    "sklearn.*",
    "xgboost.*",
    "lightgbm.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "8.0"
addopts = [
    "-ra",
    "-q", 
    "--strict-markers",
    "--strict-config",
    "--cov=mcp_trader",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-branch",
    "--cov-fail-under=70",
]
testpaths = ["src/mcp_trader/tests"]
python_classes = ["Test*", "*Test"]
python_functions = ["test_*"]
python_files = ["test_*.py", "*_test.py"]
asyncio_mode = "auto"
# Test markers
markers = [
    "unit: Unit tests",
    "integration: Integration tests", 
    "slow: Tests that take a long time to run",
    "requires_api_key: Tests that require API keys",
    "requires_network: Tests that require network access",
]

# Filter warnings
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning",
    "ignore::UserWarning",
]

[tool.coverage.run]
source = ["src/mcp_trader"]
branch = true
omit = [
    "*/tests/*",
    "*/test_*",
    "*/__init__.py",
    "*/conftest.py",
]
[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstract",
    "except ImportError:",
]
show_missing = true
skip_covered = false